aws-lightsail-alarm/src/main/java/software/amazon/lightsail/alarm/UpdateHandler.java
package software.amazon.lightsail.alarm;

import lombok.val;
import software.amazon.awssdk.services.lightsail.LightsailClient;
import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
import software.amazon.cloudformation.proxy.Logger;
import software.amazon.cloudformation.proxy.ProgressEvent;
import software.amazon.cloudformation.proxy.ProxyClient;
import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
import software.amazon.lightsail.alarm.helpers.handler.AlarmHandler;

public class UpdateHandler extends BaseHandlerStd {
    private Logger logger;

    protected AlarmHandler getAlarmHandler(final AmazonWebServicesClientProxy proxy,
                                           final ResourceHandlerRequest<ResourceModel> request, final CallbackContext callbackContext,
                                           final ProxyClient<LightsailClient> proxyClient, final Logger logger) {
        return new AlarmHandler(proxy, callbackContext, request.getDesiredResourceState(), logger,
                proxyClient, request);
    }

    protected ReadHandler getReadHandler() {
        return new ReadHandler();
    }

    protected ProgressEvent<ResourceModel, CallbackContext> handleRequest(final AmazonWebServicesClientProxy proxy,
                                                                          final ResourceHandlerRequest<ResourceModel> request, final CallbackContext callbackContext,
                                                                          final ProxyClient<LightsailClient> proxyClient, final Logger logger) {

        this.logger = logger;

        val alarmHandler = getAlarmHandler(proxy, request, callbackContext, proxyClient, logger);

        return ProgressEvent.progress(request.getDesiredResourceState(), callbackContext)
                .then(alarmHandler::handleUpdate).then(progress -> getReadHandler()
                        .handleRequest(proxy, request, callbackContext, proxyClient, logger));
    }
}
